Added a shared secret over Redis cluster.
The PR intoruduces a new shared secret that is shared over all
the nodes on the Redis cluster. The main idea is to leverage
the cluster bus to share a secret between all the nodes
such that later the nodes will be able to authenticate
using this secret and send internal commands to each other
(see #13740 for more inforamtion about internal commands).

The way the shared secret is chosen is the following:
1. Each node, when start, randomly generate its own internal
   secret.
2. Each node share its internal secret over the cluster ping messages.
3. In addition, each node also shared the secrets of the other nodes it knows about as part of the gossip protocol.
4. Node choses the secret with the minimal value (lexicographical).

The converges of the secret is as good as the topology converges.

To extend the ping messages to contain the secret, we leverage the extention machanism. Nodes that runs an older Redis version will just ignore those extentions.

Secrets are also persisted to the `node.conf` file so that on start we will have the secret right away.

Specific tests were added to verify that eventually all nodes see the secrets. In addition, a verification was added to the test infra to verify the secret on `cluster_config_consistent` and to `assert_cluster_state`